Remove physical_coordinate_calculator.py
Most of the functions are now only used by test code.  Move `_get_physical_coordinates_of_z_plane`, the remaining "mainline" code into physical_coordinates.py

Test plan: `pytest -v -n4 starfish/core/imagestack/test/`
