+++ starfish/core/imagestack/test/factories/with_coords.py
@@ -3,8 +3,8 @@ from collections import OrderedDict
 import numpy as np
 import xarray as xr
 
-from starfish.core.imagestack import physical_coordinate_calculator
 from starfish.core.imagestack.imagestack import ImageStack
+from starfish.core.imagestack.physical_coordinates import _get_physical_coordinates_of_z_plane
 from starfish.core.types import Axes, Coordinates, PhysicalCoordinateTypes
 from .synthetic_stack import synthetic_stack
 
@@ -40,9 +40,8 @@ def imagestack_with_coords_factory(stack_shape: OrderedDict, coords: OrderedDict
         np.linspace(coords[PhysicalCoordinateTypes.Y_MIN], coords[PhysicalCoordinateTypes.Y_MAX],
                     stack.xarray.sizes[Axes.Y.value]), dims=Axes.Y.value)
 
-    z_coord = physical_coordinate_calculator.\
-        get_physical_coordinates_of_z_plane((coords[PhysicalCoordinateTypes.Z_MIN],
-                                             coords[PhysicalCoordinateTypes.Z_MAX]))
+    z_coord = _get_physical_coordinates_of_z_plane(
+        (coords[PhysicalCoordinateTypes.Z_MIN], coords[PhysicalCoordinateTypes.Z_MAX]))
 
     stack.xarray[Coordinates.Z.value] = xr.DataArray(np.zeros(
         stack.xarray.sizes[Axes.ZPLANE.value]),
